backend/pyproject.toml
[project]
name = "quickt"
version = "0.1.0"
description = "QuickT — Online Bus Ticket Platform for Togo"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.1.0",
    "boto3>=1.35.0",
    "pillow>=11.0.0",
    "sentry-sdk[fastapi]>=2.0.0",
    "qrcode[pil]>=8.0",
    "python-slugify>=8.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 100
